Handle special case where auEduPersonSharedToken is invalid
If the value is too long, we're going to hit problems trying to save
it in the database, so we need to just error out and show the problem
gracefully.

If the value is otherwise invalid, it will end up being changed anyway
which might cause collisions, so it's not a great idea to snapshot it.