Branding: adjust links to attribute documentation
AAF has per-attribute documentation pages on their wiki.

As Tuakiri only has a single page with overall attributes documentation,
adjust the layout of the snapshot page and remove the fourth column with
the URL.  This can be simply done by removing the referenceURL parameter
to the template and removing the table heading.  Add instead the URL to
the overall page in the header of the listing.